requirements.txt
alembic==1.7.5
attrs==21.2.0
autopep8==1.6.0
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
iniconfig==1.1.1
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.4.0
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-flask==1.2.0
python-dotenv==0.19.2
reportlab==3.6.3
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
typer==0.4.0
Werkzeug==2.0.2
